fix bug in compute_dataset_parallel, add unit test

+++ dsmpy/modelparameters.py
@@ -39,7 +39,7 @@ class ModelParameters:
             self._n_nodes = len(radii)
             self._n_grd_params = 2 * len(radii) - 1
             n = 2 * len(radii) - 1
-            self._nodes = np.zeros(n, dtype=np.float32)
+            self._nodes = np.zeros(n, dtype=np.float64)
             for i in range(n - 1):
                 self._nodes[i] = (radii[i // 2] + (i % 2)
                                   * (radii[i // 2 + 1] - radii[i // 2]) / 2.)
@@ -308,7 +308,7 @@ class ParameterType(IntEnum):
 
 if __name__ == '__main__':
     types = [ParameterType.VSV, ParameterType.VSH]
-    radii = np.array([3480., 3700.], dtype=np.float32)
+    radii = np.array([3480., 3700.], dtype=np.float64)
     model_params = ModelParameters(types, radii)
     values_dict = {
         ParameterType.VSV: -1.,
